refactored the stock router and moved controller functions to their respective file, started moving DAL functionality from controller to service file
