+++ app/routes/stock_router.py
@@ -1,33 +1,20 @@
-from app import app, db
-from flask import jsonify, request
-from app.models import Stock, StockSchema
+from app import app
+from flask import request
+from app.controllers.stock_controller import save_stock, get_all_stocks, get_stock, delete_stock
+
+GET = 'GET'
+PUT = 'PUT'
+POST = 'POST'
+DELETE = 'DELETE'
 
 PREFIX = '/api/stock'
-stock_schema = StockSchema()
-stocks_schema = StockSchema(many=True)
 
-@app.route(PREFIX, methods=['POST'])
-def save_stock():
-  stock = Stock(title=request.json['title'], ticker=request.json['ticker'])
-  db.session.add(stock)
-  db.session.commit()
-  serialized = stock_schema.dump(stock).data
-  return jsonify({'stock': serialized}), 201
+@app.route(PREFIX, methods=[GET, POST])
+def stocks_route():
+  return save_stock() if request.method == POST else get_all_stocks()
 
-@app.route(PREFIX, methods=['GET'])
-def get_stocks():
-  stocks = Stock.query.all()
-  serialized = stocks_schema.dump(stocks).data
-  return jsonify({'stocks': serialized}), 200
+@app.route(PREFIX + '/<stock_id>', methods=[GET, DELETE])
+def stock_route(stock_id):
+  return get_stock(stock_id) if request.method == GET else delete_stock(stock_id)
 
-@app.route(PREFIX + '/<stock_id>', methods=['GET', 'DELETE'])
-def get_stock(stock_id):
-  if (request.method == 'GET'):
-    stock = Stock.query.get_or_404(stock_id, description=f'No Stock Found by ID - {stock_id}')
-    serialized = stock_schema.dump(stock).data
-    return jsonify({'stock': serialized}), 200
-  else:
-    stock = Stock.query.get_or_404(stock_id, description=f'No Stock Found By ID - {stock_id}')
-    db.session.delete(stock)
-    db.session.commit()
-    return jsonify({}), 204 
+
